working on response handler
#2
now responses are spawned at the end of a dialogue if applicable

Note:
I'm trying to follow along Semag Games' tutorial on making a flexible dialogue system. Although I'm not sure if the game will be dialogue-heavy yet, but i think I'll test it out and see how it feels.

It would also be a good script library to have for future projects

not yet implementing any aesthetic choices yet, that'll be for later