"Pass the time" puzzle
#38

It is "off" time, and you have to pass the time.
You cannot turn computer on.
You can't can only go to sleep (on mattress) between 8pm to 6am.
You will have to change the clock (a slider) to anywhere in this time range,
and go to mattress to trigger next scene.